requirements.txt
# PDF processing
PyPDF2==3.0.1
pymupdf==1.23.8

# Text processing
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.2

# Vector storage
chromadb==0.4.18
faiss-cpu==1.7.4

# Embeddings and LLM
openai==1.6.1
sentence-transformers==2.2.2

# Text preprocessing
nltk==3.8.1
spacy==3.7.2

# Web interface (optional)
streamlit==1.29.0
fastapi==0.104.1
uvicorn==0.24.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
numpy==1.24.4